api & ui. remove useless code
